Implement Basic Groundwork For Image Viewing
Through the use of a UIScrollView, we can scroll through each page of square Buttons with nested UIImages.
Button and UIImage have been extended to allow for asynchronous loading of an image at a specific URL.
Button loads in the lowest quality image in the background.
Buttons show animation when tapped.
UIScrollView is redrawn whenever the screen is rotated.